+++ ui/App.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
 import { RouterView } from "vue-router";
-import AccretionSidebar from "./components/AccretionSidebar/AccretionSidebar.vue";
 import AccretionLayout from "@/components/AccretionLayout/AccretionLayout.vue";
 import AccretionLayoutContent from "@/components/AccretionLayoutContent/AccretionLayoutContent.vue";
+import AccretionSidebar from "@/components/AccretionSidebar/AccretionSidebar.vue";
 </script>
 
 <template>