+++ ui/views/HomeView.vue
@@ -1,1 +1,14 @@
-<template>Placeholder</template>
+<script>
+import { defineComponent } from "vue";
+import AccretionSearch from "@/components/AccretionSearch/AccretionSearch.vue";
+
+export default defineComponent({
+    components: { AccretionSearch },
+});
+</script>
+
+<template>
+    <AccretionSearch />
+</template>
+
+<style lang="scss" scoped></style>